Фиксирует версии зависимостей, исправляет функцию generateDifferenceDate
