+++ package.json
@@ -10,11 +10,11 @@
     "test": "npm run eslint"
   },
   "devDependencies": {
-    "css-loader": "^5.2.7",
+    "css-loader": "5.2.7",
     "eslint": "7.31.0",
     "eslint-config-htmlacademy": "4.0.2",
     "nanoid": "3.1.22",
-    "style-loader": "^2.0.0",
+    "style-loader": "2.0.0",
     "webpack": "4.44.2",
     "webpack-cli": "3.3.12",
     "webpack-dev-server": "3.11.0"
