+++ src/mock/point-trip.js
@@ -33,8 +33,8 @@ const generatePointTrip = () => {
     id: nanoid(),
     offers: someOffers[getRandomInteger(0, someOffers.length-1)],
     destination: destinations[getRandomInteger(0, destinations.length-1)],
-    dateFrom: newDateFrom,
-    dateTo:newDateTo,
+    dateFrom: newDateFrom.format(),
+    dateTo:newDateTo.format(),
     favorite: Boolean(getRandomInteger(0, 1)),
     basePrice: getRandomInteger(0, 1000),
     availableCities: cities,
